Modify website fixture to pass in site_dir as a parameter
Originally, the test fixture changed the current working directory to
the root of the test project. This meant that if the PROJDIR indicator
(.jmdwebsite) was missing, the indicator for the main project would be
found by get_project_dir() instead. This could give false passes in
the tests.

By passing in site_dir as a parameter to the Website() constructor,
get_project_dir() is not used to find the site directory. Instead,
it is set directly, and an exceptio will be raised from within
Website() if no PROJDIR exisits in the site_dir.